requirements.txt
customtkinter
google-generativeai
python-dotenv
# Add AI libraries as needed (e.g., openai, transformers)